Refactored player online methods, added server response

+++ app/BusinessLogicLayer/managers/GameMovementManager.php
@@ -11,6 +11,7 @@ namespace App\BusinessLogicLayer\managers;
 use App\Models\api\ApiOperationResponse;
 use App\Models\GameMovement;
 use App\StorageLayer\GameMovementStorage;
+use App\Utils\ServerResponses;
 use Exception;
 
 class GameMovementManager {
@@ -45,23 +46,28 @@ class GameMovementManager {
 
     public function getLatestOpponentGameMovement($input) {
         try {
+            $playerManager = new PlayerManager();
             $gameRequestManager = new GameRequestManager();
             $gameRequest = $gameRequestManager->getGameRequest($input['game_request_id']);
+            $opponent = $gameRequest->opponent;
+            if(!$playerManager->isPlayerOnline($opponent))
+                return new ApiOperationResponse(ServerResponses::$OPPONENT_OFFLINE, 'opponent_offline', "");
+
             $gameMovement = $this->gameMovementStorage->getLatestGameMovementOfPlayer($gameRequest->id, $input['opponent_id'], $input['last_timestamp']);
             if($gameMovement) {
                 $gameMovement->delete();
-                return new ApiOperationResponse(1, 'new_movement',
+                return new ApiOperationResponse(ServerResponses::$RESPONSE_SUCCESSFUL, 'new_movement',
                     [
                         "game_movement_json" => json_decode($gameMovement->movement_json),
                         "timestamp" => $gameMovement->timestamp
                     ]);
             }
             else
-                return new ApiOperationResponse(4, 'no_movement', "");
+                return new ApiOperationResponse(ServerResponses::$RESPONSE_EMPTY, 'no_movement', "");
 
 
         } catch (Exception $e) {
-            return new ApiOperationResponse(2, 'error', $e->getMessage());
+            return new ApiOperationResponse(ServerResponses::$RESPONSE_ERROR, 'error', $e->getMessage());
         }
     }
 }
